Refinement

+++ Core/Inc/Queue.h
@@ -13,11 +13,11 @@
 #include <stdbool.h>
 
 typedef struct {
-    void* 		data;         // pointer to the data array
-    uint8_t 	head;       // starting index of the queue (0 - QUEUE_SIZE-1)
-    uint8_t 	tail;       // ending index of the queue (0 - QUEUE_SIZE-1)
-    uint8_t 	size;       // number of elements in the queue (0 - QUEUE_SIZE)
-    uint8_t		max_size;
+    void* 		data;         	// pointer to the data array
+    uint16_t 	head;       	// starting index of the queue (0 - QUEUE_SIZE-1)
+    uint16_t 	tail;       	// ending index of the queue (0 - QUEUE_SIZE-1)
+    uint16_t	size;			// to seperate the number of new items from items that we can read (for example flash save)
+    uint16_t	max_size;
     uint16_t 	item_size; // size of a signle item in queue (in bytes)
     uint32_t*	flash_page;
 } Queue;
